PassiveNarrate: read GAME_OBJECT_TYPES as 1 byte, not 4
The kind field at CGameObject +0x8 is a 1-byte enum (max 0x10 = SOUND);
reading 4 bytes captures the next 3 bytes of adjacent struct data. For
some objects those bytes are zero and the wide read works by luck —
that's why Sith-Soldat resolved fine in earlier sessions. Most objects
have non-zero trailing bytes, so wide reads return values like
0x0002000a (door, real kind=0x0a) or 0x26002e05 (creature, real
kind=0x05) that fail every kind comparison and get silenced as
"not a nav category".

Confirmed via PassiveNarrate.Probe run 2026-05-21: four distinct
handles (DOOR/CREATURE/PLACEABLE), all with valid kind bytes and
all with corrupted wide reads. The +0xf8 client→server chain was
correct in every case — only the read width was wrong.

+++ patches/Accessibility/engine_area.cpp
@@ -50,7 +50,7 @@ void* GetCurrentArea() {
 int GetObjectKind(void* gameObject) {
     if (!gameObject) return -1;
     __try {
-        return static_cast<int>(*reinterpret_cast<uint32_t*>(
+        return static_cast<int>(*reinterpret_cast<uint8_t*>(
             reinterpret_cast<unsigned char*>(gameObject) + kObjectKindOffset));
     } __except (EXCEPTION_EXECUTE_HANDLER) {
         return -1;
